+++ nodes/Hudu/descriptions/index.ts
@@ -17,6 +17,7 @@ export * from './matchers.description';
 export * from './networks.description';
 export * from './passwordFolders.descriptions';
 export * from './groups.description';
+export * from './photos.description';
 export * from './procedures.description';
 export * from './procedure_tasks.description';
 export * from './public_photos.description';