+++ nodes/Hudu/resources/index.ts
@@ -35,6 +35,8 @@ export * from './password_folders/password_folders.handler';
 export * from './password_folders/password_folders.types';
 export * from './groups/groups.handler';
 export * from './groups/groups.types';
+export * from './photos/photos.handler';
+export * from './photos/photos.types';
 export * from './procedures/procedures.handler';
 export * from './procedures/procedures.types';
 export * from './procedure_tasks/procedure_tasks.handler';